+++ pyvo/registry/tests/test_rtcons.py
@@ -363,6 +363,9 @@ class TestSelectClause:
             "res_description, "
             "reference_url, "
             "creator_seq, "
+            "created, "
+            "updated, "
+            "rights, "
             "content_type, "
             "source_format, "
             "source_value, "
@@ -371,7 +374,8 @@ class TestSelectClause:
             "\n  ivo_string_agg(COALESCE(access_url, ''), ':::py VO sep:::') AS access_urls, "
             "\n  ivo_string_agg(COALESCE(standard_id, ''), ':::py VO sep:::') AS standard_ids, "
             "\n  ivo_string_agg(COALESCE(intf_type, ''), ':::py VO sep:::') AS intf_types, "
-            "\n  ivo_string_agg(COALESCE(intf_role, ''), ':::py VO sep:::') AS intf_roles")
+            "\n  ivo_string_agg(COALESCE(intf_role, ''), ':::py VO sep:::') AS intf_roles, "
+            "alt_identifier")
 
     def test_group_by_columns(self):
         # Again, this will break as regtap.RegistryResource.expected_columns
@@ -386,8 +390,12 @@ class TestSelectClause:
             "res_description, "
             "reference_url, "
             "creator_seq, "
+            "created, "
+            "updated, "
+            "rights, "
             "content_type, "
             "source_format, "
             "source_value, "
             "region_of_regard, "
-            "waveband")
+            "waveband, "
+            "alt_identifier")